+++ Engine/Core/Regions/IRegionActivationStrategy.cs
@@ -6,6 +6,6 @@ namespace Core.Regions
     public interface IRegionActivationStrategy
     {
         HashSet<Region> GetActiveRegions();
-        void SetRegionActive(int x, int y, int z, bool active);
+        void SetRegionActive(long x, long y, long z, bool active);
     }
 }
