+++ Engine/Shared/Interfaces/IObjectFactory.cs
@@ -3,6 +3,6 @@ using Shared.Interfaces;
 namespace Shared.Interfaces;
     public interface IObjectFactory
     {
-        GameObject Create(ObjectType objectType, int x = 0, int y = 0, int z = 0);
+        GameObject Create(ObjectType objectType, long x = 0, long y = 0, long z = 0);
         void Destroy(GameObject obj);
     }
